refactor(crypto): reorganize crypto domain files and update imports
Move crypto-related domain files to dedicated crypto directory
Update import paths in affected files
Add new crypto repository and use case with tests
