+++ src/presentation/middlewares/crypto.middleware.ts
@@ -1,6 +1,6 @@
 import cryptoConfig from "@/config/crypto.config";
 import envs from "@/config/envs";
-import { DecryptPayloadUseCase } from "@/domain/use-cases/decrypt-payload.usecase";
+import { DecryptPayloadUseCase } from "@/domain/crypto/use-cases/decrypt-payload.usecase";
 import { CryptoService } from "@/infrastructure/services/crypto.service";
 import { EncryptedBodyValidator } from "@/infrastructure/validators/encrypted-body.validator";
 import { Request, Response, NextFunction } from "express";
